+++ Changelog.rst
@@ -9,6 +9,7 @@ New features
 * The ``update()`` method of a TupleHash128 and TupleHash256 hash objects
   now accepts a list of byte strings.
   Thanks to Sylvain Pelissier.
+* Added support for ECDH, with ``Crypto.Protocol.dh``.
 
 3.18.0 (18 May 2023)
 ++++++++++++++++++++++++++